+++ src/cli/commands/public-cmds/remove-cmd.js
@@ -1,9 +1,10 @@
 /** @flow */
-
+import yn from 'yn';
 import Command from '../../command';
 import { remove } from '../../../api/consumer';
 import { RemovedObjects, RemovedLocalObjects } from '../../../scope/component-remove';
 import paintRemoved from '../../templates/remove-template';
+import { removePrompt } from '../../../prompts';
 
 export default class Remove extends Command {
   name = 'remove <ids...>';
@@ -12,19 +13,38 @@ export default class Remove extends Command {
   opts = [
     ['f', 'force [boolean]', 'force remove (default = false)'],
     ['t', 'track [boolean]', 'keep tracking component (default = false)'],
-    ['d', 'delete-files [boolean]', 'delete local component files']
+    ['d', 'delete-files [boolean]', 'delete local component files'],
+    ['s', 'silent [boolean]', 'skip confirmation']
   ];
   loader = true;
   migration = true;
 
   action(
     [ids]: [string],
-    { force = false, track = false, deleteFiles = false }: { force: boolean, track: boolean, deleteFiles: boolean }
+    {
+      force = false,
+      track = false,
+      deleteFiles = false,
+      silent = false
+    }: { force: boolean, track: boolean, deleteFiles: boolean, silent: boolean }
   ): Promise<any> {
+    if (!silent) {
+      return removePrompt().then(({ shoudRemove }) => {
+        if (yn(shoudRemove)) {
+          return remove({ ids, force, track, deleteFiles });
+        }
+        return { localResult: new RemovedLocalObjects({}), remoteResult: [] };
+      });
+    }
     return remove({ ids, force, track, deleteFiles });
   }
-
-  report({ localResult, remoteResult }: { localResult: RemovedLocalObjects, remoteResult: RemovedObjects }): string {
+  report({
+    localResult = new RemovedLocalObjects({}),
+    remoteResult = []
+  }: {
+    localResult: RemovedLocalObjects,
+    remoteResult: RemovedObjects
+  }): string {
     return paintRemoved(localResult) + this.paintArray(remoteResult);
   }
   paintArray(removedObjectsArray: RemovedObjects) {