+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - support separating dev-dependencies and dev-packages from dependencies and packages when they originated from tests files 
 - support adding test files to existing component
 - ignore tracked files when running bit add and print warning message 
+- prompt user when trying to remove from remote scope
 
 ## [0.12.3] - 2018-01-28